requirements.txt
dash==4.0.0
dash-bootstrap-components==1.6.0
plotly==5.24.1
pandas==2.2.3
fredapi==0.5.2
yfinance==1.2.0
flask-caching==2.3.0
python-dotenv==1.0.1
gunicorn==23.0.0
